Factored out Namespaces class
