+++ collect_app/src/androidTest/java/org/odk/collect/android/utilities/ImageConverterTest.java
@@ -40,7 +40,7 @@ import timber.log.Timber;
 import static junit.framework.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.odk.collect.android.utilities.ApplicationConstants.XML_OPENROSA_NAMESPACE;
+import static org.odk.collect.android.utilities.ApplicationConstants.Namespaces.XML_OPENROSA_NAMESPACE;
 
 @RunWith(AndroidJUnit4.class)
 public class ImageConverterTest {
