+++ collect_app/src/main/java/org/odk/collect/android/utilities/ApplicationConstants.java
@@ -23,8 +23,6 @@ import java.util.HashMap;
 
 public class ApplicationConstants {
 
-    public static final String XML_OPENROSA_NAMESPACE = "http://openrosa.org/xforms";
-
     // based on http://www.sqlite.org/limits.html
     public static final int SQLITE_MAX_VARIABLE_NUMBER = 999;
 
@@ -104,5 +102,10 @@ public class ApplicationConstants {
         public static final int FORMS_UPLOADED_NOTIFICATION = 97;
         public static final int FORMS_DOWNLOADED_NOTIFICATION = 98;
         public static final int FORM_UPDATES_AVAILABLE_NOTIFICATION = 99;
+    }
+
+    public abstract static class Namespaces {
+        static final String XML_OPENROSA_NAMESPACE = "http://openrosa.org/xforms";
+        public static final String XML_OPENDATAKIT_NAMESPACE = "http://www.opendatakit.org/xforms";
     }
 }
