+++ collect_app/src/main/java/org/odk/collect/android/utilities/EncryptionUtils.java
@@ -64,7 +64,7 @@ import javax.crypto.spec.SecretKeySpec;
 
 import timber.log.Timber;
 
-import static org.odk.collect.android.utilities.ApplicationConstants.XML_OPENROSA_NAMESPACE;
+import static org.odk.collect.android.utilities.ApplicationConstants.Namespaces.XML_OPENROSA_NAMESPACE;
 
 /**
  * Utility class for encrypting submissions during the SaveToDiskTask.
